Split out simple operations.
